Fix change detection

+++ src/tasks/Publish/upload-package.js
@@ -1,4 +1,4 @@
-import path from "path";
+import path, { resolve } from "path";
 import execLikeShell from "./exec-like-shell";
 import getTempDir from "./get-temp-dir";
 import getGitTagName from "./get-git-tag-name";
@@ -9,34 +9,55 @@ export default async function uploadPackage(pkg, pkgPath, registry) {
   const gitpkgPackageName = getGitTagName(pkg);
   await execLikeShell("git init", pkgTempDirPkg);
   await execLikeShell("git add .", pkgTempDirPkg);
-  await execLikeShell("git commit -m gitpkg", pkgTempDirPkg);
+  await execLikeShell("git commit -m a", pkgTempDirPkg);
   await execLikeShell(`git remote add origin ${registry}`, pkgTempDirPkg);
 
   // <Relive>
+  await execLikeShell(`git fetch --all`, pkgTempDirPkg);
   await execLikeShell(`git fetch --tags`, pkgTempDirPkg);
-  let changed = await execLikeShell(
-    `git diff --quiet ${gitpkgPackageName};echo $?`,
-    pkgTempDirPkg
+
+  const { stdout } = await execLikeShell(
+    `git ls-remote --tags origin | grep refs/tags/${gitpkgPackageName}`
   );
 
-  console.log(changed);
-  changed = false;
+  if (stdout) {
+    let changed = await new Promise(async (resolve, reject) => {
+      try {
+        // Exits with code 1 when changes are available
+        await execLikeShell(
+          `git diff ${gitpkgPackageName} --quiet`,
+          pkgTempDirPkg
+        );
 
-  if (!changed) {
-    console.log("No changes detected");
-    return;
+        if (changes.code === 0) {
+          resolve(false);
+        }
+      } catch (e) {
+        if (e.code === 1) {
+          resolve(true);
+          return;
+        }
+        throw new Error(e);
+      }
+    });
+
+    if (!changed) {
+      console.log("No changes detected");
+      return;
+    }
+
+    try {
+      await execLikeShell(
+        `git push --delete origin ${gitpkgPackageName}`,
+        pkgTempDirPkg
+      );
+      await execLikeShell(`git tag -d ${gitpkgPackageName}`, pkgTempDirPkg);
+    } catch (e) {
+      console.warn(e);
+    }
   }
 
   await execLikeShell(`git tag ${gitpkgPackageName}`, pkgTempDirPkg);
-
-  try {
-    await execLikeShell(
-      `git push --delete origin ${gitpkgPackageName}`,
-      pkgTempDirPkg
-    );
-  } catch (e) {
-    console.warn(e);
-  }
   await execLikeShell(`git push origin ${gitpkgPackageName}`, pkgTempDirPkg);
   // </Relive>
 }
